# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/audit-logs": {
      "get": {
        "description": "Actions sensibles uniquement : connexions, changements de role et de statut, integrations, exports, paiements, abonnements. Les metadonnees sont expurgees de tout secret a l ecriture.",
        "operationId": "AuditController_list_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page, 1-indexee.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Taille de page. Plafonnee a 200.",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "AUTH_LOGIN_SUCCESS",
                  "AUTH_LOGIN_FAILED",
                  "AUTH_LOGOUT",
                  "AUTH_PASSWORD_RESET_REQUESTED",
                  "AUTH_PASSWORD_CHANGED",
                  "USER_INVITED",
                  "USER_ROLE_CHANGED",
                  "USER_DEACTIVATED",
                  "ROLE_PERMISSIONS_CHANGED",
                  "TENANT_CREATED",
                  "TENANT_SETTINGS_UPDATED",
                  "TENANT_SUSPENDED",
                  "ORDER_CREATED",
                  "ORDER_UPDATED",
                  "ORDER_STATUS_CHANGED",
                  "ORDER_ARCHIVED",
                  "ORDER_DUPLICATE_RESOLVED",
                  "INVENTORY_ADJUSTED",
                  "SHIPMENT_CREATED",
                  "SHIPMENT_CANCELLED",
                  "RETURN_CREATED",
                  "RETURN_RESOLVED",
                  "INTEGRATION_CONNECTED",
                  "INTEGRATION_DISCONNECTED",
                  "INTEGRATION_CONFIG_UPDATED",
                  "EXPORT_GENERATED",
                  "PAYMENT_SUBMITTED",
                  "PAYMENT_APPROVED",
                  "PAYMENT_REJECTED",
                  "SUBSCRIPTION_ACTIVATED",
                  "SUBSCRIPTION_SUSPENDED",
                  "SUBSCRIPTION_CANCELLED",
                  "TRIAL_STARTED",
                  "TRIAL_EXPIRED",
                  "TRIAL_ABUSE_FLAGGED",
                  "TRIAL_ABUSE_REVIEWED",
                  "DATA_RETENTION_APPLIED",
                  "CUSTOMER_DATA_ERASED"
                ]
              }
            }
          },
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "description": "Type d entite : Order, Payment, Membership...",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorUserId",
            "required": false,
            "in": "query",
            "description": "Auteur des actions.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Journal d audit de la boutique",
        "tags": [
          "Journal d audit"
        ]
      }
    },
    "/api/v1/admin/audit-logs": {
      "get": {
        "description": "Toutes boutiques confondues, y compris les actions sans tenant (authentification, administration).",
        "operationId": "AuditController_listGlobal_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page, 1-indexee.",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Taille de page. Plafonnee a 200.",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "AUTH_LOGIN_SUCCESS",
                  "AUTH_LOGIN_FAILED",
                  "AUTH_LOGOUT",
                  "AUTH_PASSWORD_RESET_REQUESTED",
                  "AUTH_PASSWORD_CHANGED",
                  "USER_INVITED",
                  "USER_ROLE_CHANGED",
                  "USER_DEACTIVATED",
                  "ROLE_PERMISSIONS_CHANGED",
                  "TENANT_CREATED",
                  "TENANT_SETTINGS_UPDATED",
                  "TENANT_SUSPENDED",
                  "ORDER_CREATED",
                  "ORDER_UPDATED",
                  "ORDER_STATUS_CHANGED",
                  "ORDER_ARCHIVED",
                  "ORDER_DUPLICATE_RESOLVED",
                  "INVENTORY_ADJUSTED",
                  "SHIPMENT_CREATED",
                  "SHIPMENT_CANCELLED",
                  "RETURN_CREATED",
                  "RETURN_RESOLVED",
                  "INTEGRATION_CONNECTED",
                  "INTEGRATION_DISCONNECTED",
                  "INTEGRATION_CONFIG_UPDATED",
                  "EXPORT_GENERATED",
                  "PAYMENT_SUBMITTED",
                  "PAYMENT_APPROVED",
                  "PAYMENT_REJECTED",
                  "SUBSCRIPTION_ACTIVATED",
                  "SUBSCRIPTION_SUSPENDED",
                  "SUBSCRIPTION_CANCELLED",
                  "TRIAL_STARTED",
                  "TRIAL_EXPIRED",
                  "TRIAL_ABUSE_FLAGGED",
                  "TRIAL_ABUSE_REVIEWED",
                  "DATA_RETENTION_APPLIED",
                  "CUSTOMER_DATA_ERASED"
                ]
              }
            }
          },
          {
            "name": "entityType",
            "required": false,
            "in": "query",
            "description": "Type d entite : Order, Payment, Membership...",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorUserId",
            "required": false,
            "in": "query",
            "description": "Auteur des actions.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Boutique ciblee, si l utilisateur est membre de plusieurs boutiques.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Journal d audit global (administration plateforme)",
        "tags": [
          "Journal d audit"
        ]
      }
    },
...
  },
  "info": {
    "title": "EcomFlow API",
    "description": "API de la plateforme SaaS EcomFlow — gestion et automatisation des operations e-commerce (commandes, confirmation, preparation, expedition, tracking, retours, stock, abonnement).\n\n**Multi-tenant** : chaque requete authentifiee est automatiquement limitee a la boutique de la session. L en-tete `X-Tenant-Id` permet de changer de boutique parmi celles dont l utilisateur est membre ; toute autre valeur est refusee.\n\n**Idempotence** : les endpoints de creation couteuse acceptent l en-tete `Idempotency-Key`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}